Implemented employee update pop up and set data into table
Implemented a pop-up modal for updating employee information. When an employee is selected for update, their current data is loaded into the pop-up form fields. Upon submission, the updated data is saved and the changes are reflected in the employee tables. This improves the user experience by allowing quick edits without navigating away from the table view. Also ensured that the table data is refreshed after updates to display the latest information.