Update SuperAdminController to provide accurate statistics and recent activities for Super Admin Account Management page
Added a sample set of recent activities to be displayed on the Super Admin Account Management page. This update ensures that the statistics and activity feed shown in the view are more representative of actual user data, improving the management experience for super admins. No changes were made to other controller methods.